instead of relying on the getter/setters for the jackson deserialization/serialization lets only rely on the fields + cleanup the other JacksonConfiguration classes
