--- app-ws-avam-sink/src/main/java/ch/admin/seco/jobs/services/jobadservice/infrastructure/messagebroker/avam/JacksonConfiguration.java
@@ -1,24 +1,0 @@
-package ch.admin.seco.jobs.services.jobadservice.infrastructure.messagebroker.avam;
-
-import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
-import com.fasterxml.jackson.module.afterburner.AfterburnerModule;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class JacksonConfiguration {
-
-    /*
-     * Jackson Afterburner module to speed up serialization/deserialization.
-     */
-    @Bean
-    public AfterburnerModule afterburnerModule() {
-        return new AfterburnerModule();
-    }
-
-    @Bean
-    public Jdk8Module jdk8Module() {
-        return new Jdk8Module();
-    }
-}
