+++ infrastructure/src/main/java/ch/admin/seco/jobs/services/jobadservice/infrastructure/jackson/JacksonConfiguration.java
@@ -1,0 +1,41 @@
+package ch.admin.seco.jobs.services.jobadservice.infrastructure.jackson;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.PropertyAccessor;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.module.afterburner.AfterburnerModule;
+import org.zalando.problem.ProblemModule;
+import org.zalando.problem.validation.ConstraintViolationProblemModule;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
+
+@Configuration
+class JacksonConfiguration {
+
+    @Bean
+    ObjectMapper jacksonObjectMapper(Jackson2ObjectMapperBuilder builder) {
+        ObjectMapper mapper = builder.createXmlMapper(false)
+                .build();
+        mapper.setVisibility(PropertyAccessor.ALL, JsonAutoDetect.Visibility.NONE);
+        mapper.setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.ANY);
+        return mapper;
+    }
+
+    @Bean
+    AfterburnerModule afterburnerModule() {
+        return new AfterburnerModule();
+    }
+
+    @Bean
+    ProblemModule problemModule() {
+        return new ProblemModule();
+    }
+
+    @Bean
+    ConstraintViolationProblemModule constraintViolationProblemModule() {
+        return new ConstraintViolationProblemModule();
+    }
+
+}
